# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST Skedway",
        "description": "Documentação Skedway.",
        "contact": {
            "email": "bernardodev0809@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/participant/create": {
            "post": {
                "tags": [
                    "Participante"
                ],
                "summary": "Criar um participante",
                "operationId": "27341d7a093d70429ac88c6e5d30e59f",
                "requestBody": {
                    "description": "Dados do participante",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Participante criado!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-29 00:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-29 00:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar participante!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Participante não foi criado, tente novamente mais tarde!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/participant/read-all": {
            "get": {
                "tags": [
                    "Participante"
                ],
                "summary": "Ler todos participantes",
                "operationId": "6c4922d6a79f185520dfc608c963eb70",
                "responses": {
                    "200": {
                        "description": "Participantes encontrados!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "João"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-08-29 00:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-08-29 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum participante encontrado!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Nenhum participante cadastrado até o momento!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/participant/read-by-id/{id}": {
            "get": {
                "tags": [
                    "Participante"
                ],
                "summary": "Ler participante filtrado por id",
                "operationId": "5588773920e09be8e26fd7a07cb1e1d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do participante",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Participante encontrado!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "João"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-08-29 00:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-08-29 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum participante encontrado!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Nenhum participante cadastrado até o momento!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}